+++ src/main/java/com/doublew2w/sbs/mybatis/executor/statement/PreparedStatementHandler.java
@@ -48,7 +48,6 @@ public class PreparedStatementHandler extends BaseStatementHandler {
   @Override
   public <E> List<E> query(Statement statement, ResultHandler resultHandler) throws SQLException {
     PreparedStatement ps = (PreparedStatement) statement;
-    log.info("sql:{}", boundSql.getSql());
     ps.execute();
     return resultSetHandler.<E>handleResultSets(ps);
   }