+++ src/main/java/com/doublew2w/sbs/mybatis/executor/statement/SimpleStatementHandler.java
@@ -68,7 +68,6 @@ public class SimpleStatementHandler extends BaseStatementHandler {
   @Override
   public <E> List<E> query(Statement statement, ResultHandler resultHandler) throws SQLException {
     String sql = boundSql.getSql();
-    log.info("sql:{}", sql);
     statement.execute(sql);
     return resultSetHandler.handleResultSets(statement);
   }