+++ src/main/java/com/doublew2w/sbs/mybatis/plugin/InterceptorChain.java
@@ -1,0 +1,31 @@
+package com.doublew2w.sbs.mybatis.plugin;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+/**
+ * 拦截器链
+ *
+ * @author: DoubleW2w
+ * @date: 2024/9/23 20:27
+ * @project: sbs-mybatis
+ */
+public class InterceptorChain {
+  private final List<Interceptor> interceptors = new ArrayList<>();
+
+  public Object pluginAll(Object target) {
+    for (Interceptor interceptor : interceptors) {
+      target = interceptor.plugin(target);
+    }
+    return target;
+  }
+
+  public void addInterceptor(Interceptor interceptor) {
+    interceptors.add(interceptor);
+  }
+
+  public List<Interceptor> getInterceptors() {
+    return Collections.unmodifiableList(interceptors);
+  }
+}