+++ src/main/java/com/doublew2w/sbs/mybatis/plugin/Signature.java
@@ -1,0 +1,36 @@
+package com.doublew2w.sbs.mybatis.plugin;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/9/23 20:03
+ * @project: sbs-mybatis
+ */
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({})
+public @interface Signature {
+  /**
+   * Returns the java type.
+   *
+   * @return the java type
+   */
+  Class<?> type();
+
+  /**
+   * Returns the method name.
+   *
+   * @return the method name
+   */
+  String method();
+
+  /**
+   * Returns java types for method argument.
+   * @return java types for method argument
+   */
+  Class<?>[] args();
+}