+++ src/main/java/com/doublew2w/sbs/mybatis/reflection/ExceptionUtil.java
@@ -1,0 +1,28 @@
+package com.doublew2w.sbs.mybatis.reflection;
+
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.UndeclaredThrowableException;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/9/23 20:23
+ * @project: sbs-mybatis
+ */
+public class ExceptionUtil {
+  private ExceptionUtil() {
+    // Prevent Instantiation
+  }
+
+  public static Throwable unwrapThrowable(Throwable wrapped) {
+    Throwable unwrapped = wrapped;
+    while (true) {
+      if (unwrapped instanceof InvocationTargetException) {
+        unwrapped = ((InvocationTargetException) unwrapped).getTargetException();
+      } else if (unwrapped instanceof UndeclaredThrowableException) {
+        unwrapped = ((UndeclaredThrowableException) unwrapped).getUndeclaredThrowable();
+      } else {
+        return unwrapped;
+      }
+    }
+  }
+}